+++ src/lexer/keywords/keywords.h
@@ -1,8 +1,9 @@
 #ifndef __CLOX_LEXER_KEYWORDS_H__
 #define __CLOX_LEXER_KEYWORDS_H__
 
+#include "util/common.h"
 #include "lexer/tokens.h"
 
-enum tkn_type keyword_traverse(const char *identifier, int id_len);
+enum tkn_type keyword_traverse(const char *identifier, size_t id_len);
 
 #endif // __CLOX_LEXER_KEYWORDS_H__
